+++ src/pages/Index.tsx
@@ -5,6 +5,7 @@ import Navbar from "../components/Navbar";
 import Hero from "../components/Hero";
 import Services from "../components/Services";
 import { AIChat } from "../components/AIChat";
+import { Dashboard } from "../components/Dashboard";
 import CallToAction from "../components/CallToAction";
 import TestimonialsCarousel from "../components/TestimonialsCarousel";
 import MembershipTiers from "../components/MembershipTiers";
@@ -47,6 +48,7 @@ const Index = () => {
       <Hero />
       <div className="relative z-10">
         <Services />
+        <Dashboard />
         <AIChat />
         <CallToAction />
         <TestimonialsCarousel />